fix in case the base isn't integer

+++ Demo1/main.cc
@@ -10,7 +10,7 @@
  */
 double power(double base, int exponent)
 {
-    int result = base;
+    double result = base;
     int i;
 
     if (exponent == 0) {
